requirements.txt
torch>=2.0.0
python-chess>=1.999
numpy>=1.24.0
pandas>=2.0.0
tqdm>=4.65.0
matplotlib>=3.7.0
scikit-learn>=1.3.0
